refactor: BonusLotto -> WinningLotto 변경
